# server/src/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/timesheet/getProjectTimesheetHours/:month/:year": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "month",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "year",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    }
                ]
            }
        },
        "/timesheet/getByUser/:month/:year": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "month",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "name": "year",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": ""
                    }
                ]
            }
        },
...
    },
    "swagger": "2.0"
}